Handles for resources. Tidy up of res
Handles are now passed around for a resource.
A handle is the address of the resource header.
The handle is then used to load the resource header and then the resource. This is a little slower as we are loading twice, but safe(ish) and leak free to pass to apps.